requirements.txt
flask==3.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
joblib==1.3.2
dateparser==1.2.0
